env setup

+++ src/includes/definition/type_definition.h
@@ -8,6 +8,6 @@ typedef int int32;
 typedef long long int64;
 typedef unsigned long long uint64;
 typedef float float32;
-typedef float float64;
+typedef double float64;
 
 #endif
